add physician order of e-assist study
The rds file contains the order of physicians and within each physician the number of patients. After importing the file, the order looks like 1 for physician 1 and 1 for physician 2... The file is for simulating an unbalanced dataset that has patient distribution just like e-assist study. The file doesn't contain patient or physician information.